+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.xx.x
+## 0.13.0
 
 * money question